add chat user list activity to show recent user interactions

+++ app/src/main/java/edu/northeastern/cs5520_mobileappdev_team19/ChatActivity.java
@@ -26,7 +26,7 @@ public class ChatActivity extends AppCompatActivity {
     public static final String RECIPIENT_USERNAME = "RECIPIENT_NAME";
     private String senderId;
     private String recipientId;
-    private MessagesViewAdapter<Long> messagesViewAdapter;
+    private MessagesViewAdapter<Integer> messagesViewAdapter;
     private RecyclerView messagesRecyclerView;
     private MessageService<Integer> messageService;
