+++ app/src/main/java/edu/northeastern/cs5520_mobileappdev_team19/utils/StickerMessagesViewAdapter.java
@@ -5,10 +5,9 @@ import android.content.Context;
 import androidx.annotation.NonNull;
 
 import edu.northeastern.cs5520_mobileappdev_team19.models.AbstractMessage;
-import edu.northeastern.cs5520_mobileappdev_team19.models.Sticker;
 import edu.northeastern.cs5520_mobileappdev_team19.services.StickerService;
 
-public class StickerMessagesViewAdapter extends MessagesViewAdapter<Long> {
+public class StickerMessagesViewAdapter extends MessagesViewAdapter<Integer> {
     private final StickerService stickerService;
 
     public StickerMessagesViewAdapter(Context context, String currentUserId) {
@@ -18,7 +17,7 @@ public class StickerMessagesViewAdapter extends MessagesViewAdapter<Long> {
 
     @Override
     public void onBindViewHolder(@NonNull MessagesViewHolder holder, int position) {
-        AbstractMessage<Long> stickerMessage = messages.get(position);
+        AbstractMessage<Integer> stickerMessage = messages.get(position);
         holder.messageSticker.setImageResource(stickerService.getById(Integer.parseInt(String.valueOf(stickerMessage.getContent()))).getId());
         holder.messageTimestamp.setText(stickerMessage.getTimestampAsString());
     }
